split runner.rs into multiple smaller topics
- no functional changes
